+++ CHANGELOG.md
@@ -1,10 +1,19 @@
+## 0.3.0
+
+- Added RetroAchievements::getGameInfo($gameId)
+- Added requirement for >= PHP 7.1
+- Updated unit tests
+- Improvments to RetroAchivements::__construct
+- Made returned models immutable
+- Added project contributing guidelines
+
 ## 0.2.0
 
-- Added \RetroAchievements\getConsoles()
-- Added \RetroAchievements\getGamesForConsole($consoleId)
+- Added RetroAchievements::getConsoles()
+- Added RetroAchievements::getGamesForConsole($consoleId)
 - Added PHP-CS-Fixer for PSR-2 formatting
 - Added basic Travis CI configuration
 
 ## 0.1.0
 
-- Added \RetroAchievements\getTopTenUsers()
+- Added RetroAchievements::getTopTenUsers()